extracted ask user event handler
