+++ handlers/ask_user_event_handler.py
@@ -1,0 +1,29 @@
+import keys
+from handlers.input_handlers import EventHandler
+from handlers.input_handlers import ActionOrHandler
+from typing import Optional
+
+
+import tcod
+
+
+class AskUserEventHandler(EventHandler):
+    """Handles user input for actions which require special input."""
+
+    def ev_keydown(self, event: tcod.event.KeyDown) -> Optional[ActionOrHandler]:
+        """By default any key exits this input handler."""
+        if event.sym in keys.MODIFIER_KEYS:
+            return None
+        return self.on_exit()
+
+    def ev_mousebuttondown(self, event: tcod.event.MouseButtonDown) -> Optional[ActionOrHandler]:
+        """By default any mouse click exits this input handler."""
+        return self.on_exit()
+
+    def on_exit(self) -> Optional[ActionOrHandler]:
+        """Called when the user is trying to exit or cancel an action.
+
+        By default this returns to the main event handler.
+        """
+        from handlers.main_game_event_handler import MainGameEventHandler
+        return MainGameEventHandler(self.engine)
