+++ handlers/input_handlers.py
@@ -7,7 +7,6 @@ import tcod
 from actions.action import Action
 import color
 import exceptions
-import keys
 
 if TYPE_CHECKING:
     from engine import Engine
@@ -92,26 +91,6 @@ class EventHandler(BaseEventHandler):
         self.engine.render(console)
 
 
-class AskUserEventHandler(EventHandler):
-    """Handles user input for actions which require special input."""
-
-    def ev_keydown(self, event: tcod.event.KeyDown) -> Optional[ActionOrHandler]:
-        """By default any key exits this input handler."""
-        if event.sym in keys.MODIFIER_KEYS:
-            return None
-        return self.on_exit()
-
-    def ev_mousebuttondown(self, event: tcod.event.MouseButtonDown) -> Optional[ActionOrHandler]:
-        """By default any mouse click exits this input handler."""
-        return self.on_exit()
-
-    def on_exit(self) -> Optional[ActionOrHandler]:
-        """Called when the user is trying to exit or cancel an action.
-
-        By default this returns to the main event handler.
-        """
-        from handlers.main_game_event_handler import MainGameEventHandler
-        return MainGameEventHandler(self.engine)
