+++ handlers/inventory_event_handler.py
@@ -2,7 +2,7 @@ import color
 import keys
 from engine import Engine
 from entity import Item
-from handlers.input_handlers import AskUserEventHandler
+from handlers.ask_user_event_handler import AskUserEventHandler
 from handlers.input_handlers import ActionOrHandler
 from typing import Optional
