+++ handlers/select_index_handler.py
@@ -1,7 +1,7 @@
 import color
 import keys
 from engine import Engine
-from handlers.input_handlers import AskUserEventHandler
+from handlers.ask_user_event_handler import AskUserEventHandler
 from handlers.input_handlers import ActionOrHandler
 from typing import Optional
